deploy/paddle-table/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
Pillow>=9.0.0
numpy<2
opencv-python-headless<4.11
paddleocr==2.7.3
paddlepaddle==2.6.2
